+++ Sources/Steepshot/Steepshot.Android/Assets/Localization.en-us.txt
@@ -177,5 +177,6 @@
 		"13 N5boost16exception_detail10clone_implINS0_19error_info_injectorISt12out_of_rangeEEEE: unknown key" : "unknown key",
 		"Category used for offset was not found": "Category used for offset was not found",
 		"QueryMinLength":"Query should have at least 2 characters",
+		"4100000 plugin_exception: plugin exception: The comment is archived":"The comment is archived",
     }
 }
