+++ Sources/Steepshot/Steepshot.Android/Utils/AssetsHelper.cs
@@ -27,9 +27,14 @@ namespace Steepshot.Utils
             return TryReadAsset<ConfigInfo>("Config.txt");
         }
 
-        public LocalizationModel GeLocalization()
+        public LocalizationModel GetLocalization(string lang)
         {
-            return TryReadAsset<LocalizationModel>("Localization.txt");
+            return TryReadAsset<LocalizationModel>($"Localization.{lang}.txt");
+        }
+
+        public void SetLocalization(LocalizationModel model)
+        {
+            TryWriteAsset($"Localization.{model.Lang}.txt", model);
         }
 
         public HashSet<string> TryReadCensoredWords()
@@ -77,6 +82,24 @@ namespace Steepshot.Utils
                 //to do nothing
             }
             return new T();
+        }
+
+        private void TryWriteAsset<T>(string file, T data)
+        {
+            try
+            {
+                var json = JsonConvert.SerializeObject(data);
+                var stream = _assetManager.Open(file);
+                using (var sr = new StreamWriter(stream))
+                {
+                    sr.Write(json);
+                }
+                stream.Dispose();
+            }
+            catch
+            {
+                //to do nothing
+            }
         }
     }
 }
