+++ Sources/Steepshot/Steepshot.Core/Localization/LocalizationKeys.cs
@@ -1,0 +1,168 @@
+﻿namespace Steepshot.Core.Localization
+{
+    public enum LocalizationKeys
+    {
+        WrongPrivatePostingKey,
+        WrongPrivateActimeKey,
+        EmptyResponseContent,
+        ResponseContentContainsHtml,
+        UnexpectedUrlFormat,
+        EnableConnectToServer,
+        EnableConnectToBlockchain,
+        ServeNotRespond,
+        ServeUnexpectedError,
+        EmptyCommentField,
+        UnknownError,
+        UnknownCriticalError,
+        EmptyTitleField,
+        EmptyBodyField,
+        EmptyCategoryField,
+        EmptyPhotoField,
+        EmptyFileField,
+        EmptyContentType,
+        EmptyVerifyTransaction,
+        EmptyUrlField,
+        EmptyUsernameField,
+        EmptyLogin,
+        EmptyPostingKey,
+        EmptyActiveKey,
+        EmptyPostPermlink,
+        EmptyCategory,
+        PhotoProcessingError,
+        PhotoPrepareError,
+        PhotoUploadError,
+        ErrorCameraPreview,
+        ErrorCameraScale,
+        ErrorCameraZoom,
+        FollowError,
+        PostTagsError,
+        InternetUnavailable,
+        IncorrectIdentifier,
+        MaxVoteChanges,
+        UnexpectedError,
+        CameraSettingError,
+        VotedInASimilarWay,
+        TagLimitError,
+        UnsupportedMime,
+        UnexpectedProfileData,
+        PostFirstComment,
+        ScanQRCode,
+        Comments,
+        PostSettings,
+        RapidPosting,
+        CameraHoldUp,
+        WaitforScan,
+        Likes,
+        Like,
+        Flags,
+        Flag,
+        Follow,
+        Unfollow,
+        Error,
+        Ok,
+        TryAgain,
+        Forget,
+        Voters,
+        FlagVoters,
+        ViewComments,
+        FlagPhoto,
+        HidePhoto,
+        Feed,
+        Trending,
+        Hot,
+        New,
+        Top,
+        Clear,
+        SeeComment,
+        Login,
+        Photos,
+        Following,
+        Followers,
+        Reply,
+        NewPhotos,
+        Hello,
+        Profile,
+        AcceptToS,
+        ChoosePhoto,
+        TypeTag,
+        TypeUsername,
+        YourAccountName,
+        NextStep,
+        Account,
+        NsfwShow,
+        Nsfw,
+        NsfwContent,
+        NsfwContentExplanation,
+        LowRated,
+        LowRatedContent,
+        LowRatedContentExplanation,
+        FlagMessage,
+        FlagSubMessage,
+        DeleteAlertTitle,
+        DeleteAlertMessage,
+        PowerOfLike,
+        TitleForAcceptToS,
+        PostDelay,
+        SignInButtonText,
+        CreateButtonText,
+        EnterAccountText,
+        PasswordViewTitleText,
+        PublishButtonText,
+        AppSettingsTitle,
+        AddAccountText,
+        PeopleText,
+        TapToSearch,
+        SearchHint,
+        Tag,
+        Users,
+        YearsAgo,
+        MonthAgo,
+        DaysAgo,
+        DayAgo,
+        HrsAgo,
+        HrAgo,
+        MinAgo,
+        SecAgo,
+        CreateFirstPostText,
+        Copied,
+        PostLink,
+        ShowMoreString,
+        SignIn,
+        FlagPost,
+        UnFlagPost,
+        FlagComment,
+        HideComment,
+        EditComment,
+        DeleteComment,
+        UnFlagComment,
+        HidePost,
+        EditPost,
+        DeletePost,
+        CopyLink,
+        Cancel,
+        Delete,
+        PutYourComment,
+        ServeRejectRequest,
+        LoginMsg,
+        NoAccountMsg,
+        AppVersion,
+        AppVersion2,
+        BadRequest,
+        Forbidden,
+        NotFound,
+        InternalServerError,
+        BadGateway,
+        GatewayTimeout,
+        HttpErrorCode,
+        QueryMinLength,
+        EnterPostTitle,
+        EnterPostDescription,
+        AddHashtag,
+        MyProfile,
+        AccountBalance,
+        ShowNsfw,
+        ShowLowRated,
+        Guidelines,
+        ToS,
+    }
+}
